Issue #694 Fix contract between copier and key
We would leak a non copied key inside functions when configuring
copyForRead. This was an issue as it could be observed from JSR-107
entry processor.
Added tests to guarantee copier contract at the key level.
One downside is the increase in serialization / deserialization
happening when interacting with a key serializing on heap store.