Issue #165 Handle eviction events properly in offheap based stores
Eviction in offheap happens before the real put / replace operation can
take place as the memory limit is hard. In this context, we need the
eviction event to be part of the same context as the mutation event
to prevent deadlocks.
It also requires some event rewriting in case the mutation and
eviction happens on the same key.