Issue #811 Correct faulty ServiceLocator test
The org.ehcache.core.spi.ServiceLocatorTest.testCircularDeps test had a
couple of problems rendering the test ineffective:

* @Mock was used to "instantiate" the services used in the test.  The
  generated mock classes do not carry the @ServiceDependencies
  annotation so the dependencies were not actually observed by
  ServiceLocator.  In addition, the mock classes do not carry the expected
  concrete class name so the instances actually registered in the
  ServiceLocator are not the same as the names declared in the
  ServiceDependencies annotations.
* The assertThat clause used in the start methods of the test services
  was coded using the Mockito.notNull method instead of
  Matchers.notNullValue method.  This resulted in the test actually being
  for null instead of not null.  Because of the @Mock interactions (above)
  the faulty test incorrectly passed.

This commit replaces the mocks with formal test classes and the test
classes are changed to include tracking of the state formerly checked
using Mockito methods.

This commit also updates testServicesInstanciatedOnceAndStartedOnce --
this test includes a check for constructor calls to the
TestProvidedService used -- this concrete class implementing this
interface is also used in other tests so its constructor may be called
multiple times.