Issue #811 Support registration of multiple implementations of a Service subtype
This commit introduces the PluralService annotation with which Service subtypes
may be marked to indicate that multiple instances of that type can be registered
in ServiceLocator.  Initially, Store.Provider, AuthoritativeTier, and CachingTier
are marked with PluralService.

In addition to the primary work in this commit, the CacheProvider interface is
moved to the test source branch -- it is not used in production code.