Issue #1108 Prevent client stall on CacheManager.close
This commit changes the behavior of DefaultClusteringService.close to
avoid Entity calls that might result in an interaction with the server.
Server interactions made when the server is disconnected can result in
operation stalls which are not resolved until the server returns to
service.  This kind of stall makes little sense when closing a
CacheManager.

This commit also adds an integration-test test case to reproduce the
symptoms of the issue.  TestLoggingService is added to enable
client-side logging from integration-test unit tests.