Implement Ehcache operation timeouts
This commit introduces operation timeouts to Ehcache for control of the
time clustered operations take.  One timeout is exposed through
configuration -- a GET operation timeout.  The other timeouts are either
internal or experimental.

In general, EhcacheClientEntity is changed to support timeouts for get,
mutative, and lifecycle operations.  When a timeout is recognized, a
TimeoutException is thrown from the EhcacheClientEntity.  (This commit
also removes a stall related to replicated operations.)

ServerStore methods, including those of ServerStoreProxy and its
subclasses, re-throw a TimeoutException caught from EhcacheClientEntity.

The ClusteredStore implementation is changed to throw a
StoreAccessTimeoutException when a TimeoutException is raised for a
mutative operation.  The StoreAccessTimeoutException is a
and RuntimeException bypasses the ResilienceStrategy in place for the
cache.  For get operations, ClusteredStore returns a 'null'.

Lifecycle operations performed by EhcacheClientEntityFactory and
EhcacheClientEntity are changed to throw a TimeoutException.  Changes
to DefaultClusteringService rethrow the TimeoutException as a
RuntimeException or another exception appropriate for the operation.