Issue #1158 Review modifications
This commit makes changes based on review comments.  The most
significant of which is the removal of the
ClusteringServiceClientSideConfiguration interface and
implementation and use ClusteringServiceConfigurationBuilder
instead.  This introduces yet another constructor to
ClusteringServiceConfiguration.

In addition to the ClusteringServiceConfigurationBuilder changes, code
in StrongServerStoreProxy is changed to use
EhcacheClientEntity.invokeAsync instead of EhcacheClientEntity.invoke.