Close #1282 Add close() method down ServerStoreImpl chain.
During rapid creation/destruction of caches/cache managers,
I observed that GC did not cleanup the dangling offheap resources
leading to extremely high memory consumption and an eventual
failure to allocate more buffers.

The theory was that the object was to complex for easy GC'ing in the
direct buffer allocation path. By closeing the server store, we
eventually remove the reference to each of the ChainMaps, which then
makes it easy for the GC to clean up the memory.

Plus PR fixups