Conditionally disable TerminatedServerTest concurrency
Test concurrency is disabled in TerminatedServerTest if the
'disable.concurrent.tests' Java system property is set.  build.gradle
is changed to set this property to 'true' if running on CloudBees.
The concurrency level for non-CloudBees environments is changed
to the lesser of 1/2 the number of tests and the number of available
processors.

CloudBees detection is done by checking for 'cloudbees' in the
JENKINS_URL environment variable.  Setting JENKINS_URL to
'cloudbees' can be used to inhibit concurrent test execution in
non-CloudBees environments.